requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
twikit==2.1.3
pydantic==2.6.1
python-dotenv==1.0.0 